Remove unnecessary live updating previews broadcast (#116)
This code was refreshing every `iframe` on the page, which was wreaking
havoc on the WebContainer boot process (since it uses an iframe with the
proprietary stackblitz code).

It uses a BroadcastChannel to signal across tabs whether it should
reload iframes. This is incidentally unnecessary since we have HMR for
our Vite server.

I've reverted this file to be just the bolt.new OSS code.

This seems to improve WebContainer boot stability dramatically!