Add path for provisioning projects for normal auth flow (#131)
The user logs in with their Convex credentials, and they get a team
picker on the initial screen (teams are loaded from big brain using
their Convex credentials).

Chats automatically create a new project in that team by piping through
the Convex credentials to Convex function calls and ultimately
exchanging them for a project key which we store and use for operating
on the deployment in flex.

We removed a couple things which we'll add back in subsequent PRs --
disconnecting + reconnecting a project, better error handling
(especially around hitting the project limit), but want to get this in
now.

---------

Co-authored-by: Ari <ari@convex.dev>