Add a access gate screen between signing in + using chef (#160)
Re-using the invite codes, but not for their intended purpose. Now folks
who sign in will get a screen asking for an invite code. If they've
previously entered a valid code, they will skip this step.

In this case, we're not using the invite codes as user identity (that's
coming from the auth with convex.dev), but instead using it as a basic
UI gate.