Simplify the rendering tree for homepage vs. chat routes (#165)
This PR removes the SSR path for `BaseChat`, since we're committed to
showing a spinner for auth anyways.

This then makes the path of route -> Chat -> FlexAuthWrapper -> ChatImpl
-> BaseChat a bit simpler, since each component only has a single parent
(other than `Chat`, which is used in both routes)

This is prep for doing more synchronous data loading before showing the
UI in the `/chat` route.