Add support for amazon bedrock + retries for convex agent (#132)
I have learned so much about the Vercel AI SDK.

After doing some research, I have found that it is not possible to retry
requests if `streamText` returns an error. The headers of the response
are not exposed, and there is no way to `await` them. Below are some
relevant threads:
- https://github.com/vercel/ai/discussions/3387
- https://github.com/vercel/ai/issues/4842

I also learned that the `streamText` function already retries 3 times on
failures. I opted to turn this off in favor of client-side retries
because it is not possible to switch between providers on the retry.

I have opted to do client side retries. This means that if the user
sends a request and it fails, we will give them a prompt to retry again.
When we retry, we flip between the providers (bedrock vs. anthropic). If
this fails three times in a row, we tell the user that Chef is down and
to check back later.

In a follow-up PR, I will add the ability for users to enter their own
API key.

This is the error that users will see when their request fails.
![Screenshot 2025-04-08 at 2 48
38 PM](https://github.com/user-attachments/assets/0015ce42-76bd-4d78-a5e0-8103807f10ce)

Before landing this in prod, I need to set the `USE_ANTHROPIC_FRACTION`
env variable.

I tested the retry logic locally and everything works as expected.

Fixes ENG-8821 ENG-8820