convert partial-call to failed state on initialMessages load (#203)
When there was a partial-call in the messages history on reload, the UI
would be unusable because the partial-call would make the state stuck in
"Cooking..." with spinners on the artifact for the partial call. This
changes the `useInitialMessages` hook to make these appear as "failed".
It doesn't write back to the db bc that also made it harder to
repro/test and I'm haven't 100% thought through any weird implications
of writing the change back... but I think this is good enough to prevent
the unusable loading state.