Allow user-specified API keys (#204)
Adds ability for users to enter their own api key. Currently, we only
support Anthropic API keys and will only use Claude 3.5 Sonnet.

An short overview of the data model/flow is that:
- api keys are stored next to the member object in the `convexMembers`
table
- api keys can only be set for the current logged in user
- the api key is threaded through the request to vercel
- in vercel we check usage and decide whether or not to use the token
based on the user's preference

This was tested locally to make sure it works as expected.

Fixes ENG-8848