Restore “Add Preview” feature (#202)
This fixes an issue where app auth would get disconnected when the agent
deploys functions. I do this by avoiding restarting the dev server if
it's already running. The fix itself is in
9be872e61d807ec4de255eeddfb29153221aed2c

With @emmaling27 I also made the Convex deploy action that runs when
editing files manually run in its dedicated terminal.

---------

Co-authored-by: Emma Forman Ling <emma@convex.dev>