Prompt for auth after initial message gets sent (#225)
We open our own page at `/signin` and make you sign in with github.

Opening an Auth0 popup directly doesn't work because I think they
require that the auth domain (auth.convex.dev) is the same as the
requesting domain (chef.convex.dev).

In our mini window, we go through the flow and set the relevant state in
local storage, which the original tab is subscribed to.

The other change is I stopped automatically considering the first team
in the list of fetched teams the default choice (unless there is only
one team).

(I think there are some auth0 settings I need to tweak)