Add PostHog analytics (#268)
Adds PostHog to `root.tsx`, following a simplified version of what we
use on the website, Stack, Docs, etc. For a quick start it uses
cookieless tracking only, but we can later add a cookie banner as we do
on the other sites.

Sample of events (when `debug: true` is set):

![Screenshot 2025-04-10 at 2 04
23 PM](https://github.com/user-attachments/assets/51bb53fd-a0c7-4534-b332-ce8683e061ce)

This is the most basic set up for now, and once we confirm it's working
we can always add custom events or anything else needed to make it more
useful.