GPT-4.1 Prompting (#322)
The default behavior of this model is to:
- ask for lots of user input (would originally ask about a plan without
writing any code)
- often terminate its execution before deploying code

The model behavior after prompting is that:
- it will now write the front and backend functions after the first user
prompt
- will always typecheck its code on subsequent user messages

The biggest gap in this model's behavior is its ability to deploy the
code after creating the files after the first shot. I tried for a few
hours to get this working, but for some reason the model refuses to do
it. We can work more on this in the coming days, but I'm not sure how to
message this to our customers.

However, with sufficiently telling this model to continue on, it does
create working apps with UIs that are generally better than claude.

I mostly included prompt changes to `openAi.ts`, but included a few
minor things that are shared with other models since they will probably
also be helpful for them.