Make GPT-4.1 better at deploying changes (#335)
I read
[this](https://cookbook.openai.com/examples/gpt4-1_prompting_guide)
article and learned that for GPT-4.1, important prompts perform better
if they are at both the top and the bottom.

This worked really well! Now the model remembers to deploy changes at
the end of its first turn and fix any errors. It would be sick if we
could get this out today.