Auto-retry on first api request failure (#323)
Added behavior that we will automatically resend the request on the
first model call fail. This will allow us to immediately switch
providers in case we are getting rate-limited.

I tested this locally and confirmed that it works.

Once this lands in prod, we can turn on sending all traffic to anthropic
first.

Initially, the plan was to do exponential backoff, but turns out its
quite tricky to do this from the client, and we also added a `resend`
button which should make the UX a lot better in case our model calls are
failing.

Fixes ENG-9061

---------

Co-authored-by: Sujay Jayakar <sujayakar@convex.dev>