Fix HMR (#355)
I can't figure out how to convince the web container to let me run the
Vite HTTP server and the Vite WS on different ports: it's clear in the
Vite docs but the the WebSocket was getting a 400 back.

So instead expand the proxy to proxy WebSocket connections as well.