React perf (#393)
Pass the parsed messages through directly to BaseChat component instead
of building new message objects and a new message array for it. This
makes the message objects retain their identities between renders
(great!) since parsed messages have this property. It also ties UI
render speed to the speed at which we parse messages, controlled by
`processSampledMesages` sampler.

This dataflow is wacky with processSampledMessages in a useEffect, it'd
be better to combine this with useMessageParser.
And tying "processing" of the messages (interpreting them, this has side
effects) and the update rate at which we render them is less flexible,
but at 50ms the existing value is fine for UI too.

Also fix a few small regressions causes over-rendering by memoizing.
Also kill the hooks barrel file.