Add Grok 3 Mini (#368)
Tool use is broken right now for Grok 3, so we can either add it later
or support both Grok 3 and Grok 3 mini.

The apps this model makes isn't great and will probably require some
additional prompting to get it great, but we can do that separately.

I didn't gate this change because I think we're just happy to release
as-is and improve prompting later.