Fix duplicate artifacts rendering (#392)
This would happen when the LLM would return stuff like
```
<boltArtifact>
   <boltAction id='fileA'>....</boltAction>
</boltArtifact>
<boltArtifact>
   <boltAction id='fileB'>....</boltAction>
</boltArtifact>
```
in a single part instead of one artifact wrapping multiple actions.

We store action + artifact state keyed on the message ID + part ID, so
we happen to merge the state of all the artifacts together so the
actions get executed as expected, but on the UI side we try and render
the artifact associated with this part N times.

As a bit of a hack, let's just include the special tag for the artifact
once per part to make it only render once