Persist message history when a part finishes (#359)
This is easy for tool call parts (`state` is `result`), but less
straightforward for text parts. A text part should be finished if either
the next part has started (e.g. a `text` part followed by a tool call)
or we stop streaming (`streamStatus` is something other than
`streaming`).

I believe the `streamStatus` and the `messages` should always be
consistent with each other (there's `processSampledMessage`, but I
believe it'll keep arguments at a consistent snapshot).

We can switch this to persisting an immutable list of message histories
instead but I have not done that in this diff yet -- that's up next.