Add migration for message history (#434)
I'd like to migrate all our chat messages to one format to make our
lives easier.

The top of `messageMigration` describes the two migrations I want to
perform.

I want to get this code in now, but I think the most I'll do to start is
to first dry run some of the share migrations, then do a few for real.

In general, I'm worried that there are corner cases in prod that I'm
currently omitting -- I think my migration does nothing when it hits a
state it does not understand vs. potentially writing incorrect data, so
I might have to loop back and edit these before running the full
migration.

(also this should go in with
https://github.com/get-convex/flex-diy/pull/428)