Rewind (#363)
* Makes the `chatMessagesStorageState` table a log indexed by `[chatId',
'lastMessageRank', 'partIndex']`. We store the chat state and filesystem
state (new field - `snapshotId`) there. We stop updating the
`snapshotId` field in `chats` table.
* Rewinding means setting the `lastMessageRank` in `chats` table to the
message you want to rewind to. We can extend this in the future to
include the part of the message. We keep around the
`chatMessagesStorageState` to support restoring to a future checkpoint,
until you send another message (then the later `lastMessageRank`s get
deleted).
* The UI here shows a rewind button on each message - there's some
improvements to be made that @sshader and I have discussed. Maybe we
start with just showing the rewind button for the latest user message as
a stepping stone until we have better UI that would support helpful
hints if there's big code and data skew from going back many
checkpoints.
* I added lots of tests based on bugs I ran into.

---------

Co-authored-by: Sarah Shader <sarahshader@gmail.com>