Hide the absolute URL from the UI in `<Preview>` (#473)
https://github.com/user-attachments/assets/0978860c-4b83-4da9-8332-57ad8b648467

This hides the WebContainers origin, which was confusing to users
because they would just assume you could use them outside of Chef.
(Bolt.new does the same thing)

Known issue: if the user navigates from within the preview to another
URL, the URL bar won’t update. I don’t think there is a simple way to
fix this since for security reasons the `<iframe>` parent doesn’t have
access to the iframe URL (but maybe we could in the future inject a
script in the preview that can `postMessage` when the URL change).
Bolt.new has the same limitation (and the dashboard tab on Chef has it
too)