Record annotations with the chat's initial ID instead of the less unique urlId (#479)
The initial ID is a UUID, so we should be able to reliably look up more
information about the chat in our Chef backend. The URL ID, which we
were preferring before, can be something like `chat-app`, which is only
unique within a session.

Am trying to avoid adding session IDs in our axiom logging since they're
somewhat security sensitive.