Automatically stop the agent if it's failed many times in a row (#476)
I have a theory that the agent will just repeatedly keep trying to do
things even when there's a problem it doesn't quite know how to fix
(e.g. I delete the auth.ts and it low key is incapable of recovering).

Instead of burning through token usage here, I made it stop after
several consecutive failed calls.

I maintain state on the frontend of # of consecutive failed calls, and
pipe that down in the `/chat` call and use that to disable tools. I
tried to condition the prompts to summarize the failed tool calls to the
user and ask them what they want to do next, but couldn't get that
working, so I manually have a little indicator on the frontend
conditioned on an annotation