Fail on consecutive deploy failures, not just consecutive tool call failures (#512)
I've seen the agent do things like fail to deploy 4 times, `npm install`
something successfully, and the fail to deploy 4 more times, and I think
whenever it fails to deploy many times in a row, we should just fail.

Also if we add more granular edit tools, I don't think succeeding to
edit should reset the count here.

Something I've also noticed but have not addressed in this PR -- we'll
charge for the last turn of the agent loop on failure, which is a little
different than when we'd just let the agent loop on indefinitely