Don't apply Anthropic-specific caching headers unconditionally (#526)
If there's no giant system prompt, don't try to cache anything. This
saves cache creation tokens for us. Not a huge deal (they're only 1.25x
the cost of normal prompt tokens) but we didn't intend for this to
happen.

A related issue where we double-count cache creation tokens made this
more noticeable.