Add prompt around conditional hooks (#530)
I saw this `Rendered more hooks than during the previous render.` error
many times when trying the slack sample prompt, which is due to
conditional hooks, which is a bad react pattern. I added a specific
prompt to fix this and tested it.

I also saw an error where chef repeatedly doesn't know how to access the
current user. I added an example for this and tested it.