Fix issue with anonymous users (#557)
Turns out that the issue with not seeing messages was related to being
an anonymous user. So, I added a prompt to make UIs better for anonymous
users. Essentially, the messages wouldn't render if the user didn't have
a profile name.

Tested this with like 5 slacks locally and the other prompts and things
are working well.