Prevent workbench from covering scrollbar (#598)
The workbench/preview currently has margin and padding that increase its
width, capturing scroll events and blocking the user from clicking the
scrollbar for the underlying page.


https://github.com/user-attachments/assets/595448ea-fdf2-4725-bf68-baf1b6b5f8e5

(Note the tiny area above the preview where the scrollbar can be
clicked)

This PR reduces the margin and padding on the workbench, and adjusts the
`--workbench-inner-width` CSS variable to keep the workbench the same
size, while ensuring there isn't a transparent area covering the
scrollbar.


https://github.com/user-attachments/assets/5ef4da0e-3177-400a-96c3-d2de643e1b07