message input box styling (#599)
Refreshes the style of <MessageInput> and alerts next to it. Separating
the action area visually, making the contents of the action bar smaller,
adding shadows, and making the scrollbar in the text area prettier


https://github.com/user-attachments/assets/a1573a8f-00f3-4d60-936d-af70760ebf66



----

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.