Put relevant files in user messages (#595)
This PR puts relevant files in user messages when they are sent from the
chat so that sending new messages doesn't result in cache misses for
user messages. We wrap the relevant files in `<boltArtifact>` tags,
similar to how user modifications are wrapped. This prevents the UI from
rendering them. This approach is extensible to rendering which files are
sent up and allowing the user to choose which files to send up.

In a follow up, I want to experiment with only sending relevant files if
they are not already in previous messages. Right now this is redundant
if the agent just wrote to a file and we send it right back up. It's a
bigger model change so want to split it out into its own PR.