Don't include relevant files when resending a message (#603)
Now that we're sending relevant files with user messages, if the model
fails and resends, it doubles the number of tokens used. This PR makes
sure we don't send relevant files if we are resending a message. I'll
clean this up in a follow up where we'll be more selective about when we
send relevant files, but want to get this out.