Only send relevant files if they are not in the chat history context (#605)
Relevant files are a bunch of tokens and now that we send more of the
chat history with each request, we can probably get away with only
sending relevant files when they're not present in the chat history.
We'll refresh them when we collapse messages. Gets rid of checking if
this is a resent message since checking for relevant files covers that
case.