Fix reload bug (#613)
The `messages` variable was not up to date when the `onError` callback
was executed, os we were restarting the assistant message inadvertently.
There isn't a simple way to solve this in react, so I opted to remove
the automatic retries.

This means that when the model hits an error, the user has to resend
their message.

Fixes ENG-9194