Don't exclude the relevant files in serialization (#616)
We used to want to exclude the file contents when storing chat history,
but now we rely on it for context since we don't send up relevant files
separately in the chat. This means folks are probably encountering bad
performance after reloading the page because no file contents are
getting sent.