Fix repeated failed tool calls (#644)
After a tool call fails, the system would only be able to make one more
deploy before failing again. This was happening because we were just
counting the total # of consecutive failed deploys even if they were in
different messages. This change simplifies the code by only checking the
current message for failed deploys. if the most recent assistant message
has more than the allowed failed deploys, then we show disable the
tools.

I tested this locally to confirm that this works as expected.