Terminate Braintrust on failure (#654)
Currently Braintrust will `exit(0)` when the eval fails to compile (e.g.
https://github.com/get-convex/chef/actions/runs/15195895582 which shows
a green tick while there is a module loading error). I think that we’re
supposed to use `--terminate-on-failure` here? But I’m not 100% sure,
since this could also mean that we’re exiting when the eval compiles but
then the LLM doesn’t output a valid result 🤔

```
  --terminate-on-failure
                        If provided, terminates on a failing eval, instead of the default (moving onto the next one).
```