Update `presence` component guidelines (#696)
Make it such that we don't read the `userId` field that is passed into
the component. Instead, we just fetch the authenticated user since we
always have auth in chef.

This prevents the LLM passing bad values in for `userId` to cause the
component to not work.

This consistently works a lot better than the old example.