Add unsupported components to system prompt (#700)
This adds a list of unsupported components to the system prompt so chef
doesn't try to use them, but most of the time chef still tries to use
them, and the `lookupDocs` toolcall fails, and then it says the
component is unsupported. I think that's good enough.

Follow up idea is to add a trace to see what components folks want to
use so we can prioritize well.