[New Feature Button] Update data model + add migrations (#712)
Updates the data model to properly support the new fields described in
[this](https://www.notion.so/convex-dev/Chef-New-Feature-Button-20ab57ff32ab80f286f6e38be16edbf3?d=20eb57ff32ab8059b7cb001c7dbfe293#20eb57ff32ab806b8fe7f00dc0c30887)
doc. We now write `subchatIndex` and `lastSubchatIndex` in all relevant
chats.

I updated tests and confirmed that things work in dev as expected. I had
to also update the `shares` table to include a `subchatIndex` since it
is at a point in time (added this in the design doc too).

I will run the migration to backfill all the new fields once these
changes land in staging and then prod.