Skip writing new chatMessagesStorageState documents for message parts (#719)
We currently write documents (and files) to `chatMessagesStorageState`
on every message part update. We originally wanted to be able to rewind
to message parts, but we never built that and I don't think we need to
so we can stop writing all this extra data and patch the document
instead, deleting the old snapshots as we go.