Delete old chatMessagesStorageState docs and files for the same lastMessageRank (#728)
This PR adds a migration that deletes documents in the
`chatMessagesStorageState` table that are not the latest for the
`lastMessageRank` (documents for earlier message parts). Also deletes
the filesystem snapshot and chat history files for those old chat
states. This is cleanup for write behavior we already implemented in
#719

The function that does all the deletes is
`deleteOldStorageStatesForLastMessageRank`. We paginate over the `chats`
table, then for each chat, we paginate over `chatMessagesStorageState`,
and for each `lastMessageRank`, we delete documents and files. The
pagination and scheduling is to avoid hitting transaction limits - some
of the chats have a lot of messages, and some of the messages have a lot
of parts.

I tested on my dev deployment and it cut file storage in half.