Delete orphaned files migration (#732)
This PR changes the `deleteAllOldChatStorageStates` migration to not
delete any files and just delete documents in
`chatMessagesStorageState`, and adds a new migration
`deleteOrphanedFiles` that walks the storage table and deletes files
that are not referenced.

I think this approach is better because we encountered a few bugs when
testing the old approach. It turns out it is really tricky to know what
snapshot ids will be referenced until after you delete the documents in
`chatMessagesStorageState`. I think the issue of deleting snapshot ids
that are still referenced could easily follow us in the next migration
to remove the `chatMessagesStorageState` documents beyond a certain
rewind threshold. This new approach is reusable - we don't have to write
any custom logic to find references, we will just run the migration to
clean up the state first, then clean up any orphaned files later, for
both deleting the unnecessary part storage states and the states beyond
the rewind threshold.