[New Feature Button] Add API for creating new subchats (#734)
Added all the backend functionality to support creating new subchats +
relevant tests. I created a rudimentary UI that works with these new
endpoints to confirm that these endpoints are all that's necessary to
build a UI (not included in this PR).

This PR does not include updates to the share + rewind flows, which will
be updated in a following PR.