make earliestRewindableMessageRank a reactive query (#749)
We'll need this when we start deleting storage states. Making the query
reactive will prevent us from rendering a rewind button when there is no
storage state to rewind to.

This changes `earliestRewindableMessageRank` to be a reactive query so
that as we delete storage states, it updates automatically and the
rewind button disappears from messages that you can no longer rewind to.
Before, we called a one-shot query (not reactive) and passed
the`earliestRewindableMessageRank` down the react component tree as a
number. Now it's a query.