Rewind threshold (30 messages) and migration to delete old chat storage states (#746)
This PR adds a migration to delete `chatMessagesStorageState` documents
and unreferenced files that are more than 30 messages before the latest
message and schedules a mutation to do the same in `updateStorageState`.
Now that we've cleaned up all the non-latest duplicate documents for
`lastMessageRank` parts, I'm not worried about hitting transaction
limits. The deletion logic is in `deleteOldChatStorageStatesForChat`,
and we call this fn from the migration and schedule it in
`updateStorageState`.