[New Chats] Make shares work with multiple subchats (#752)
Copies the latest storage state for each subchat index. This is okay
because once a subchat is finished, you can't go back and send more
messages in it.

For point-in-time shares, this copies all previous subchat indexes and
then uses the state specified in the `shares` table for the current
subchat.

I tested this locally and added tests.