Fix file edit tool failing on relative path (#757)
I've seen file edit tool calls fail because it's using the relative path
instead of the absolute path. This PR changes behavior from throwing an
error on relative path to checking if it's relative or absolute and
returning the relative path, since that's what's used anyway.

Doesn't fix a known issue with file edit toolcalls getting called in a
loop.