Charge on failed toolcalls (#762)
This deletes the code where we skip recording usage for failed
toolcalls. We've fixed the failed toolcall loops in #761 so I think this
is ok to deploy. There's a lot more code cleanup associated with this,
but I just cleaned up the token usage tracking UI for now, and will
follow up later.