[New Chat] Implement UI (#748)
This PR does the heavy lifting for creating the new subchat UI. There
are a few tweaks to state management because you can now switch between
chats. The main pieces to review here are the `SubchatBar` and changes
to the `ChatSyncWorker`.

Please play around with the preview a bit. Shares and rewinds should
work. The UI changes are gated behind a feature flag, so these changes
shouldn't affect users yet.

There are still a few UI tweaks I'd like to make in follow-ups like:
- showing subchats in a dropdown
- auto-summarization of previous features
- a nudge to create a new subchat after the messages get sufficiently
long