[New Chat] Allow shares when a subchat is empty (#781)
When a new subchat is created, it is possible that the `storageId` is
`null`. This caused an error even though this is a valid state. We now
allow the chat history to be empty if we are on a subchat that is after
the first one.

I tested locally and added a test to confirm that this works.