[New Chat] Fix chat sync worker bug (#785)
Adds a `subchatIndex` field to the `ChatSyncState` and updates it
whenever we fetch messages from the server so we can always know which
messages are loaded in the `ChatSyncWorker`. Previously, the loaded
messages could be out of sync with the selected subchat index which
caused bugs when navigating between different subchats.

Also adds a hook called `useIsSubchatLoaded` that checks if the sync
state contains the messages from the selected subchat. This allows us to
provide better handling for loading states around the app.

There are some other minor changes in here including:
- splitting out the stores for subchat related things into its own file
- adding nice loading states when navigating between different subchats
- loading subchats directly inside of the `useInitialMessages` hook,
that way we don't have to update `initialMessages` for the `subchats`
list to update