add resend component (#772)
Adds support for using the resend component, behind a feature flag. The
agent will prompt the user to input required environment variables
(`RESEND_API_KEY`, `RESEND_DOMAIN`, and `RESEND_WEBHOOK_SECRET`) and set
up the webhook in resend. Once those steps are done, the generated app
works.

We still need to get rid of the resend proxy integration and instruct
the agent to always use the resend component when sending emails. That's
not in this PR, but this should be enough to give to @mikecann and
partners to test out as long as they put "use the resend component to
send emails" in the prompt.

![Screenshot 2025-07-07 at 11 04
40 AM](https://github.com/user-attachments/assets/5b4db938-1832-4a0d-ad31-920cadfc68ee)