[Bugfix] Don't start generation until convex project is successfully created (#799)
We were previously seeing issues where the convex project would not be
successfully created, but the LLM would continue generated. This would
cause an issue because the user would lose their chat history.

We now block generating on a project being successfully created, and if
it fails, we show an error to the user.

Fixes ENG-9501