[Bugfix] Reset context manager state (#803)
When a user creates a switches subchats, we should reset the context
manager state. This can cause us to not collapse the messages correctly
and cause the user to to run into errors like
[this](https://vercel.com/convex-dev/chef/9WjeT3wFnqJWVRgLRfSctdeZzNtn/logs?selectedLogId=27rx5-1752884905917-df4555fae57b&refreshedAt=1752885191760).
This happens when we send up an empty set of messages because the
indexes are calibrated incorrectly.

I have been able to repro this in staging, so I will push to it and then
test.