Use oauth client id and secret instead of arbitrary app name (#805)
The `appName` parameter to the `/authorize` endpoint is being
deprecated. We can use the client id an secret for the Chef oauth app
(there's one for development, staging, and production) instead.

I updated the environment variables on the production and staging
deployment. Developers will need to resync the default project env vars