[API Keys] Allow pro users to use built-in chef tokens when api key is not provided (#809)
We were getting some complaints on
[discord](https://discord.com/channels/1019350475847499849/1399469900631642123)
that users with other API keys set were not able to use the built-in
chef tokens when API keys were set for another provider.

This fixes the issue by checking if the user has an API key set for the
current provider they have selected instead of any provider.