update auth to use workos (#833)
Updates Chef to use WorkOS instead of Auth0, and all the changes needed
to make that happen. The largest change is that we can no longer query
members by `tokenIdentifier`, so I've taken the liberty of backfilling a
new column in called `convexMemberId`, which exists on the WorkOS JWT.

However, this data doesn't exist on the auth0 jwt - so the database
state may go out of sync in the time between when I ran the last
migration, and when this PR makes it to prod. Meaning that some small
set of new users won't be able to log in until I run the backfill
migration again. I will run the migration again right before deploying
this to prod, and right after deploying to prod as well.

Other notable changes:
- The login flow no longer opens a new window. It does a redirect flow
similar to the dashboard.