[Open Source] Update readme (#850)
Moves the development instructions for `README.md` to `DEVELOPMENT.md`.
We want to do this because most people coming to this repo don't need to
know about it. Instead, I added a brief description of what Chef is and
added an overview of the repo.

We can tweak/change the copy later. This is more of a structural change.