Clean up anthropic caching (#855)
Since [this](https://www.anthropic.com/news/token-saving-updates)
released, we only need to add a cache point at the end of the messages
we send.

This allows us to remove the hack we did earlier. I tested locally and
confirmed that we still are putting/reading tokens from the cache.