[Open Source] Lock down prompt enhancement (#880)
Previously, this didn't require any type of auth and anyone could get a
ton of inference by hitting this endpoint. This adds auth so users that
are over the limit/not authed don't get responses.

I tested this locally to confirm it works. This will break existing
users that try to enhance a prompt on the old chef version, but they can
just reload the page.